requirements.txt
enum34==1.0.4
nose==1.3.7
peewee==2.6.3
psycopg2==2.6.1
pycrypto==2.6.1
Twisted==15.4.0
zope.interface==4.1.2
